fix: moved edits into access denied tsx
