+++ apps/journeys-admin/src/components/AccessDenied/AccessDenied.tsx
@@ -1,3 +1,4 @@
+import { gql, useMutation} from '@apollo/client'
 import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
 import Divider from '@mui/material/Divider'
@@ -5,24 +6,37 @@ import List from '@mui/material/List'
 import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
 import NextLink from 'next/link'
-import { ReactElement } from 'react'
+import router from 'next/router'
+import { ReactElement, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import ChevronLeftIcon from '@core/shared/ui/icons/ChevronLeft'
 import HelpCircleContainedIcon from '@core/shared/ui/icons/HelpCircleContained'
 import Lock1Icon from '@core/shared/ui/icons/Lock1'
 
+import { UserJourneyRequest } from '../../../__generated__/UserJourneyRequest'
+
 import { AccessDeniedListItem } from './AccessDeniedListItem'
 
-interface AccessDeniedProps {
-  handleClick?: () => void
-  requestedAccess?: boolean
-}
+export const USER_JOURNEY_REQUEST = gql`
+  mutation UserJourneyRequest($journeyId: ID!) {
+    userJourneyRequest(journeyId: $journeyId, idType: databaseId) {
+      id
+    }
+  }
+`
 
-export function AccessDenied({
-  handleClick,
-  requestedAccess = false
-}: AccessDeniedProps): ReactElement {
+export function AccessDenied(): ReactElement {
+  const [userJourneyRequest] =
+    useMutation<UserJourneyRequest>(USER_JOURNEY_REQUEST)
+  const [requestedAccess, setRequestReceived] = useState(false)
+
+  const handleClick = async (): Promise<void> => {
+    await userJourneyRequest({
+      variables: { journeyId: router.query.journeyId }
+    })
+    setRequestReceived(true)
+  }
   const { t } = useTranslation('apps-journeys-admin')
 
   return (
